third_party/WebKit/LayoutTests/external/wpt/css/css-masking/parsing/mask-border-outset-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Masking Module Level 1: parsing mask-border-outset with valid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.fxtf.org/css-masking-1/#the-mask-border-outset">
<meta name="assert" content="mask-border-outset supports the full grammar '[ <length> | <number> ]{1,4}'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("mask-border-outset", "10px");
test_valid_value("mask-border-outset", "10px 20");
test_valid_value("mask-border-outset", "10px 20 -30px");
test_valid_value("mask-border-outset", "10px 20 -30px -40");
</script>
</body>
</html>
